requirements.txt
Pillow
arrow==0.14.2
aws-lambda-builders==0.3.0
aws-sam-cli==0.18.0
aws-sam-translator==1.10.0
awscli==1.16.201
binaryornot==0.4.4
boto==2.49.0
boto3==1.9.191
botocore==1.12.191
certifi==2019.6.16
chardet==3.0.4
chevron==0.13.1
click==6.7
colorama==0.3.9
cookiecutter==1.6.0
dateparser==0.7.1
docker==3.7.3
docker-pycreds==0.4.0
docutils==0.14
Flask==1.0.4
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.1.1
poyo==0.4.2
pyasn1==0.4.5
pymongo==3.7.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
regex==2019.6.8
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
serverlessrepo==0.1.8
six==1.11.0
tzlocal==1.5.1
urllib3==1.24.1
websocket-client==0.56.0
Werkzeug==0.15.5
whichcraft==0.6.0
